+++ packages/core/src/crypto/index.js
@@ -5,3 +5,4 @@ export * from './counterwallet.js';
 export * from './hd.js';
 export * from './wif.js';
 export * from './walletBlob.js';
+export * from './backup.js';